fds quero dormir passei o update para o main

+++ app/src/main/java/com/example/projetopdm/MainActivity.java
@@ -5,17 +5,23 @@ import static com.example.projetopdm.LocalDataBase.FuncionarioSharedPreferences.
 
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.constraintlayout.widget.ConstraintLayout;
+import androidx.exifinterface.media.ExifInterface;
 import androidx.room.Room;
 
 import android.app.Activity;
+import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.graphics.Matrix;
 import android.net.ConnectivityManager;
 import android.net.NetworkInfo;
+import android.net.Uri;
 import android.os.Bundle;
 import android.util.Base64;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -32,13 +38,20 @@ import com.example.projetopdm.BackEnd.RetrofitClient;
 import com.example.projetopdm.LocalDataBase.AppDatabase;
 import com.example.projetopdm.LocalDataBase.DAOs.NotaRMADao;
 import com.example.projetopdm.LocalDataBase.DAOs.RMADao;
+import com.example.projetopdm.LocalDataBase.Entity.NotaRMAEntity;
 import com.example.projetopdm.LocalDataBase.Entity.RMAEntity;
 import com.example.projetopdm.Modelos.Funcionario;
+import com.example.projetopdm.Modelos.NotaRMA;
 import com.example.projetopdm.Modelos.RMA;
 import com.example.projetopdm.databinding.ActivityMainBinding;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
 
+import java.io.ByteArrayOutputStream;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -59,6 +72,7 @@ public class MainActivity extends AppCompatActivity {
     Api api;
     AppDatabase db;
     RetrofitClient retrofitClient;
+    private boolean isConnectedPreviously = false;
 
     ArrayList<RMA> rmaList = new ArrayList<RMA>();
     ListAdapterRMA listAdapter;
@@ -79,6 +93,11 @@ public class MainActivity extends AppCompatActivity {
     }
 
 
+
+
+
+
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -88,10 +107,12 @@ public class MainActivity extends AppCompatActivity {
         loading.setVisibility(View.VISIBLE);
         initializeFuncionarioFromIntent();
         initializeDatabaseAndViewModel();
+
         setupPerfilButton();
         displayFuncionarioImage();
 
         setContentView(binding.getRoot());
+
 
         ImageView perfil_btn = findViewById(R.id.perfil_btn);
         searchView = findViewById(R.id.searchView);
@@ -124,8 +145,186 @@ public class MainActivity extends AppCompatActivity {
             }
         });
 
+        justDoIT();
+
 
     }
+
+    private void justDoIT() {
+        ArrayList<RMAEntity> rmasModificados = new ArrayList<>();
+        if(notaRMADao.getAllNotasRMA()!=null){
+            for (NotaRMAEntity x:notaRMADao.getAllNotasRMA()) {
+                if(x.getOffSync()!=null){
+                    if (x.getOffSync().equals("modificado")||x.getOffSync().equals("novo")||x.getOffSync().equals("apagado")){
+                        if(!rmasModificados.contains(rmaDao.getRMAById(x.getRMAId()))){
+                            rmasModificados.add(rmaDao.getRMAById(x.getRMAId()));
+                        }
+                    }
+                }
+
+            }
+        }
+
+        if (rmasModificados.size()!=0){
+            updateBaseDeDados();
+        }
+
+    }
+
+    public void updateBaseDeDados(){
+        Log.e("Notas","aqui gayyyyyyyyy");
+
+        ArrayList<NotaRMA> novos= new ArrayList<>();
+        ArrayList<NotaRMA> modificados = new ArrayList<>();
+        ArrayList<NotaRMA> apagados = new ArrayList<>();
+
+        for (NotaRMAEntity x : notaRMADao.getAllNotasRMA()) {
+            if (x.getOffSync()!=null){
+                if (x.getOffSync().equals("novo")){
+                    novos.add(x.toNotaRMA());
+                } else if (x.getOffSync().equals("modificado")) {
+                    modificados.add(x.toNotaRMA());
+                } else if (x.getOffSync().equals("apagado")) {
+                    apagados.add(x.toNotaRMA());
+                }
+            }
+        }
+        for (NotaRMA x:novos) {
+            String request ="";
+            Log.e("Notas","teste img  "+x.getImagemNota());
+            if (x.getImagemNota() != null) {
+                Uri uri = Uri.parse(x.getImagemNota());
+                Bitmap imagem = uriToBitmap(getApplicationContext(), uri);
+                String imagemString = bitmapToString(imagem);
+                x.setImagemNota(imagemString);
+                request = "{"
+                        + " \"Id\": \"" + 0 + "\", "
+                        + " \"Titulo\": \"" + x.getTitulo() + "\", "
+                        + " \"Nota\": \"" + x.getNota() + "\", "
+                        + " \"RMAId\": \"" + x.getRMAId() + "\", "
+                        + " \"IdImagem\": \"" + 0 + "\", "
+                        + " \"Imagem\": \"" + imagemString + "\" }";
+            } else {
+                request = "{"
+                        + " \"Id\": \"" + 0 + "\", "
+                        + " \"Titulo\": \"" + x.getTitulo() + "\", "
+                        + " \"Nota\": \"" + x.getNota() + "\", "
+                        + " \"RMAId\": \"" + x.getRMAId() + "\", "
+                        + " \"IdImagem\": \"" + 0 + "\", "
+                        + " \"Imagem\": \"" + "" + "\" }";
+            }
+
+            JsonObject body = new JsonParser().parse(request).getAsJsonObject();
+            Call<JsonObject> call = RetrofitClient.getInstance().getMyApi().CreateOrUpdateNotaRMA(body);
+
+            call.enqueue(new Callback<JsonObject>() {
+                @Override
+                public void onResponse(Call<JsonObject> call, retrofit2.Response<JsonObject> response) {
+                    JsonObject responseObj = response.body().get("Result").getAsJsonObject();
+                    if (responseObj.get("Success").getAsBoolean()) {
+                        Toast.makeText(getApplicationContext(), "Nota criada com sucesso", Toast.LENGTH_LONG).show();
+                        Intent resultIntent = new Intent();
+                        resultIntent.putExtra("AtivarAPI", true);
+                        setResult(Activity.RESULT_OK, resultIntent);
+
+                        notaRMADao.deleteById(x.getId());
+                        int id= response.body().get("NotaRMAId").getAsInt();
+                        x.setId(id);
+                        notaRMADao.insert(x.toNotaRMAEntity());
+
+
+
+
+                        //teste apagar depois
+
+
+
+
+                    } else {
+                        Toast.makeText(getApplicationContext(), "Erro ao criar nota", Toast.LENGTH_LONG).show();
+                    }
+                }
+
+                @Override
+                public void onFailure(Call<JsonObject> call, Throwable t) {
+                    Toast.makeText(getApplicationContext(), "Erro ao criar nota", Toast.LENGTH_LONG).show();
+                }
+            });
+
+        }//passar os novos da local para a online
+        for (NotaRMA x:modificados){
+            String request ="";
+
+            if (x.getImagemNota() != null) {
+                Uri uri = Uri.parse(x.getImagemNota());
+                Bitmap imagem = uriToBitmap(getApplicationContext(), uri);
+                String imagemString = bitmapToString(imagem);
+                x.setImagemNota(imagemString);
+            }
+            request = "{"
+                    + " \"Id\": \"" + x.getId() + "\", "
+                    + " \"Titulo\": \"" + x.getTitulo() + "\", "
+                    + " \"Nota\": \"" + x.getNota() + "\", "
+                    + " \"RMAId\": \"" + x.getRMAId() + "\", "
+                    + " \"IdImagem\": \"" + x.getImagemNotaId() + "\", "
+                    + " \"Imagem\": \"" + x.getImagemNota() + "\" }";
+
+            JsonObject body = new JsonParser().parse(request).getAsJsonObject();
+            Call<JsonObject> call = RetrofitClient.getInstance().getMyApi().CreateOrUpdateNotaRMA(body);
+
+            call.enqueue(new Callback<JsonObject>() {
+                @Override
+                public void onResponse(Call<JsonObject> call, retrofit2.Response<JsonObject> response) {
+                    JsonObject responseObj = response.body().get("Result").getAsJsonObject();
+                    if (responseObj.get("Success").getAsBoolean()) {
+                        Toast.makeText(getApplicationContext(), "Nota alterada com sucesso", Toast.LENGTH_LONG).show();
+                        Intent resultIntent = new Intent();
+                        resultIntent.putExtra("AtivarAPI", true);
+                        setResult(Activity.RESULT_OK, resultIntent);
+                        notaRMADao.deleteById(x.getId());
+                    } else {
+                        Toast.makeText(getApplicationContext(), "Erro ao alterar nota", Toast.LENGTH_LONG).show();
+                    }
+                }
+
+                @Override
+                public void onFailure(Call<JsonObject> call, Throwable t) {
+                    Toast.makeText(getApplicationContext(), "Erro ao alterar nota", Toast.LENGTH_LONG).show();
+                }
+            });
+
+        }
+        for (NotaRMA x:apagados){
+
+            Call<JsonObject> call = RetrofitClient.getInstance().getMyApi().DeleteNotaRMA(x.getId());
+
+            call.enqueue(new Callback<JsonObject>() {
+                @Override
+                public void onResponse(Call<JsonObject> call, retrofit2.Response<JsonObject> response) {
+                    JsonObject responseObj = response.body().get("Result").getAsJsonObject();
+                    if (responseObj.get("Success").getAsBoolean()) {
+                        Toast.makeText(getApplicationContext(), "Nota apagada com sucesso", Toast.LENGTH_LONG).show();
+                        Intent resultIntent = new Intent();
+                        resultIntent.putExtra("AtivarAPI", true);
+                        setResult(Activity.RESULT_OK, resultIntent);
+                        notaRMADao.deleteById(x.getId());
+                    } else {
+                        Toast.makeText(getApplicationContext(), "Erro ao apagar nota", Toast.LENGTH_LONG).show();
+                    }
+                }
+
+                @Override
+                public void onFailure(Call<JsonObject> call, Throwable t) {
+                    Toast.makeText(getApplicationContext(), "Erro ao apagar nota", Toast.LENGTH_LONG).show();
+                }
+            });
+
+        }
+
+    }
+
+
+
     private void initializeFuncionarioFromIntent() {
         Funcionario x = getFuncionarioData(this);
 
@@ -141,10 +340,13 @@ public class MainActivity extends AppCompatActivity {
 
         // Inicialização do objeto Funcionario com os dados obtidos
         funcionario = new Funcionario(id, guid, nome, email, contacto, pin, imagemFuncionario, estadoFuncionarioId, estadoFuncionario);
-        if(!funcionario.getGUID().equals(x.getGUID())){
-            this.deleteDatabase("BaseDeDadosLocal");
-            saveFuncionarioData(this,funcionario);
+        if (x!=null){
+            if(!funcionario.getGUID().equals(x.getGUID())){
+                this.deleteDatabase("BaseDeDadosLocal");
+                saveFuncionarioData(this,funcionario);
+            }
         }
+
     }
     private void initializeDatabaseAndViewModel() {
 
@@ -314,5 +516,65 @@ public class MainActivity extends AppCompatActivity {
             return null;
         }
     }
+    public Bitmap uriToBitmap(Context context, Uri uri) {
+        Bitmap bitmap = null;
+        try {
+            // Abre um InputStream a partir da URI
+            InputStream imageStream = context.getContentResolver().openInputStream(uri);
+
+            // Obtém a rotação da imagem a partir das informações EXIF
+            int rotation = getRotationFromExif(context, uri);
+
+            // Converte o InputStream em um Bitmap considerando a rotação
+            bitmap = BitmapFactory.decodeStream(imageStream);
+            if (rotation != 0) {
+                Matrix matrix = new Matrix();
+                matrix.postRotate(rotation);
+                bitmap = Bitmap.createBitmap(bitmap, 0, 0, bitmap.getWidth(), bitmap.getHeight(), matrix, true);
+            }
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        return bitmap;
+    }
+    private int getRotationFromExif(Context context, Uri uri) {
+        int rotation = 0;
+        try {
+            InputStream input = context.getContentResolver().openInputStream(uri);
+            ExifInterface exifInterface = new ExifInterface(input);
+
+            int orientation = exifInterface.getAttributeInt(ExifInterface.TAG_ORIENTATION, ExifInterface.ORIENTATION_NORMAL);
+
+            switch (orientation) {
+                case ExifInterface.ORIENTATION_ROTATE_90:
+                    rotation = 90;
+                    break;
+                case ExifInterface.ORIENTATION_ROTATE_180:
+                    rotation = 180;
+                    break;
+                case ExifInterface.ORIENTATION_ROTATE_270:
+                    rotation = 270;
+                    break;
+            }
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        return rotation;
+    }
+    public String bitmapToString(Bitmap bitmap) {
+        ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream();
+        // Comprime a imagem em um formato específico (PNG, JPEG, etc.)
+        bitmap.compress(Bitmap.CompressFormat.PNG, 100, byteArrayOutputStream);
+        byte[] byteArray = byteArrayOutputStream.toByteArray();
+
+        // Converte os bytes para Base64
+        String encodedImage = Base64.encodeToString(byteArray, Base64.DEFAULT);
+        return encodedImage;
+    }
+
 
 }
